refactor: 요구사항 재해석 반영 할인 정책이 아닌 할인 정보
