--- src/main/java/coupon/domain/DiscountPolicy.java
@@ -1,5 +1,0 @@
-package coupon.domain;
-
-public interface DiscountPolicy {
-    int apply(int money);
-}
