+++ src/test/java/coupon/domain/DiscountMoneyTest.java
@@ -1,0 +1,30 @@
+package coupon.domain;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class DiscountMoneyTest {
+
+    @DisplayName("할인 금액은 1,000원 이상이어야 한다.")
+    @Test
+    void validateMinimumDiscountMoney() {
+        assertThatThrownBy(() -> new DiscountMoney(500))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("할인 금액은 10,000원 이하여야 한다.")
+    @Test
+    void validateMaximumDiscountMoney() {
+        assertThatThrownBy(() -> new DiscountMoney(10500))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("할인 금액은 500원 단위로 설정할 수 있다.")
+    @Test
+    void validateDiscountUnit() {
+        assertThatThrownBy(() -> new DiscountMoney(100))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+}
